Split out extension cache from async RustConnection
This commit moves the code for managing information about supported X11
extensions in x11rb-async from rust_connection/mod.rs to
rust_connection/extensions.rs. The goal here is to make that file
smaller and hopefully less complex.
